removed Index and IndexMut implementations, fixed printing
as they leaked the implementation detail of saving a 0 for empty cells
the formatting function prints _ instead of 0 for empty cells

the SudokuSolver struct was moved into the sudoku module to access private
sudoku fields, as no public interface exists at this point

added comments for the solver
