+++ src/sudoku.rs
@@ -1,10 +1,10 @@
 use consts::*;
 use entry::Entry;
 use parse_error::ParseError;
-use solver::SudokuSolver;
 
-use std::{fmt, ops, slice, iter};
+use std::{fmt, slice, iter};
 use std::io::BufRead;
+use covers::Covers;
 
 #[derive(PartialEq, Eq, Debug, Clone)]
 /// The main structure exposing all the functionality of the library
@@ -103,45 +103,165 @@ fn num_to_opt(num: &u8) -> Option<u8> {
 
 impl fmt::Display for Sudoku {
 	fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-		for entry in self.iter().enumerate().map(|(cell, num)| Entry { cell: cell, num: num.unwrap()} ) {
+		for entry in self.0.iter().enumerate().map(|(cell, &num)| Entry { cell: cell, num: num } ) {
 			try!( match (entry.row(), entry.col()) {
 				(_, 3) | (_, 6) => write!(f, " "),    // seperate fields in columns
 				(3, 0) | (6, 0) => write!(f, "\n\n"), // separate fields in rows
 				(_, 0)          => write!(f, "\n"),   // separate lines not between fields
 				_ => Ok(()),
 			});
-			try!(write!(f, "{}", entry.num()));
+			//try!(
+            try!( match entry.num() {
+                0 => write!(f, "_"),
+                1...9 => write!(f, "{}", entry.num()),
+                _ => unreachable!(),
+            });
+                //uwrite!(f, "{}", entry.num())
+            //);
 		}
 		Ok(())
 	}
 }
 
-// Indexing by cell number, numbering starts at 0
-impl ops::Index<usize> for Sudoku {
-	type Output = u8;
+// Solving happens by an exact cover algorithm
+// There are a total of 729 (81 cells * 9 numbers) sudoku entry possibilities
 
-	fn index(&self, idx: usize) -> &Self::Output {
-		self.0.index(idx)
-	}
-}
-
-impl ops::IndexMut<usize> for Sudoku {
-	fn index_mut(&mut self, idx: usize) -> &mut Self::Output {
-		self.0.index_mut(idx)
-	}
-}
-
-// row/column-indexing
-//impl ops::Index<(usize, usize)> for Sudoku {
-//    type Output = u8;
+// every entry (cell-number-combination) satisfies 4 constraints
+// 1. a row    needs to have 1 of each number (9 rows, 9 numbers each)
+// 2. a column needs to have 1 of each number (9 cols, 9 numbers each)
+// 3. a field  needs to have 1 of each number (9 fields, 9 numbers each)
+// 4. a cell needs to be filled               (81 cells, 1 number each)
 //
-//	fn index(&self, (row, col): (usize, usize)) -> &Self::Output {
-//		self.0.index(row*9+col)
-//	}
-//}
+// for a total of 81*4 = 324 constraints
+//
+// the covers property in SudokuSolver contains the information what entries can
+// be added at a certain point in the solving process, which constraints are
+// already satisfied and how many possibilities still exist for a given constraint.
+// See also the covers module.
+//
+// Solving happens by recursively walking the tree of possible sudokus
+// If some constraint can only be satisfied by 1 entry, it will be entered immediately
+// This is equivalent to finding naked singles and hidden singles
+// If no entry can be deduced, a constraint with the least amount of possibilites
+// is chosen and all possibilites tried out.
 
-//impl ops::IndexMut<(usize, usize)> for Sudoku {
-//	fn index_mut(&mut self, (row, col): (usize, usize)) -> &mut Self::Output {
-//		self.0.index_mut(row*9+col)
-//	}
-//}
+// helper struct for recursive solving
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct SudokuSolver {
+	pub grid: Sudoku,
+	pub covers: Covers,
+}
+
+impl SudokuSolver {
+	pub fn from_sudoku(sudoku: Sudoku) -> SudokuSolver {
+		SudokuSolver {
+			covers: Covers::from_sudoku(&sudoku),
+			grid: sudoku,
+		}
+	}
+
+	#[inline]
+	fn remove_impossible(&mut self, new_entry: Entry) {
+		self.covers.remove_impossible(new_entry);
+	}
+
+	fn insert_entry(&mut self, entry: Entry) {
+		self.grid.0[entry.cell()] = entry.num();
+		self.remove_impossible(entry);
+	}
+
+	fn with_entry(&self, entry: Entry) -> Self {
+		let mut sudoku = self.clone();
+		sudoku.insert_entry(entry);
+		sudoku
+	}
+
+	#[inline]
+	pub fn is_solved(&self) -> bool {
+		&self.covers.covered[..] == &[true; 324][..]
+	}
+
+	#[inline]
+	fn is_impossible(&self) -> bool {
+		Iterator::zip( self.covers.possibilities_count.iter(), self.covers.covered.iter() )
+			.any(|(&poss, &covered)| !covered && poss == 0)
+	}
+
+	// return true if new entries were found
+	fn insert_deduced_entries(&mut self) -> bool {
+		let entries = self.covers.possibilities_count.iter()
+			.enumerate()
+			.filter(|&(_, &n_poss)| n_poss == 1)
+			.map(|(idx, _)| self.matching_entry(idx) )//	self.covers.entries.iter().cloned().find(|e| e.constrains(idx))
+			.collect::<Vec<_>>();
+
+		let entries_added = entries.len() != 0;
+		for entry in entries {
+			self.insert_entry(entry);
+		}
+		entries_added
+	}
+
+	// may fail, but only if used incorrectly
+	#[inline]
+	fn matching_entry(&self, constraint_nr: usize) -> Entry {
+		self.covers.entries.iter()
+			.cloned()
+			.find(|e| e.constrains(constraint_nr))
+			.unwrap()
+	}
+
+	pub fn solve_one(self) -> Option<Sudoku> {
+		let result = self.solve_at_most(1);
+		if result.len() == 0 {
+			None
+		} else {
+			result.into_iter().next() // just take one
+		}
+	}
+
+	pub fn solve_unique(self) -> Option<Sudoku> {
+		let result = self.solve_at_most(2);
+		if result.len() == 1 {
+			result.into_iter().next()
+		} else {
+			None
+		}
+	}
+
+	pub fn solve_at_most(self, limit: usize) -> Vec<Sudoku> {
+		let mut solutions = vec![];
+		self._solve_at_most(limit, &mut solutions);
+		solutions
+	}
+
+	fn _solve_at_most(mut self, limit: usize, solutions: &mut Vec<Sudoku>) {
+		if solutions.len() == limit { return }
+
+		// deduce entries, but check in between deductions if the sudoku is still possible
+		while self.insert_deduced_entries() {
+			if self.is_impossible() { return }
+		}
+
+		// impossible to insert another number
+		// either solved or unsolvable
+		if self.covers.is_empty() && self.is_solved() {
+			if self.is_solved() {
+				solutions.push(self.grid);
+			}
+			return
+		}
+
+		let (idx, _) = self.covers.possibilities_count.iter()
+			.enumerate()
+			.filter(|&(_, &n_poss)| n_poss != 0)
+			.min_by_key(|&(_, n_poss)| n_poss)
+			.unwrap();
+		for trial_sudoku in self.covers.entries.iter()
+			.filter(|e| e.constrains(idx))
+			.map(|&new_entry| self.with_entry(new_entry))
+		{
+			trial_sudoku._solve_at_most(limit, solutions);
+		}
+	}
+}
